NEW QR CODE SYSTEM & NEW OBJECT TYPE SYSTEM
If you want to use QR codes, you have to run new canopy
You have to work with Charlie S. who has to update Canopy configs to match your installation

The "TreeOrShrub" variable has been changed to a Type enum
There is a new system of "pieces" which are now named.
A pieceId is the human readable string for a part of the sculpture (like a tree or a shrub or anything else)

This code seems to work end to end with Canopy on a branch Charlie told me about and he should merge in shortly.

NOTE: the current system might not be performant enough. I've put the pieceId on every Cube, which is
a lot of string compares, and the right answer is to use "pieceIndex" which I invented later when
I started worrying. That will be an internal improvement though - this works OK for now.